Eliminated duplicate class for fetching environment variables and system properties
